requirements.txt
python-telegram-bot==21.7
Flask==3.0.3
python-dotenv==1.0.1
Flask-CORS==4.0.1
gunicorn==21.2.0
